SdkError and profiler moved out to tools-common; internal-api-clients no longer depends on common
