+++ common-libs/common/test/unit/index.test.ts
@@ -3,4 +3,3 @@ import 'mocha'
 require('./_baseDocumentLoader.test')
 require('./Affinity.test')
 require('./services')
-require('./shared/profiler/ProfilerDecorator.test')
