+++ common-libs/internal-api-clients/package.json
@@ -43,7 +43,7 @@
   "homepage": "https://github.com/affinityproject/affinidi-core-sdk#readme",
   "dependencies": {
     "@affinidi/affinidi-did-auth-lib": "^1.0.8",
-    "@affinidi/common": "^2.0.0-beta.6",
+    "@affinidi/tools-common": "^1.0.0-beta.1",
     "lodash.keyby": "^4.6.0",
     "node-fetch": "^2.6.0"
   },
