+++ common-libs/internal-api-clients/src/services/DidAuthApiService.ts
@@ -1,4 +1,4 @@
-import { profile } from '@affinidi/common'
+import { profile } from '@affinidi/tools-common'
 
 import { RawApiSpec, RequestOptionsForOperation } from '../types/request'
 import { BuiltApiType } from '../types/typeBuilder'
