+++ common-libs/internal-api-clients/src/services/KeyStorageApiService.ts
@@ -1,4 +1,4 @@
-import { profile } from '@affinidi/common'
+import { profile } from '@affinidi/tools-common'
 
 import keyStorageSpec from '../spec/_keyStorage'
 import { ParseSpec } from '../types/openapiParser'
