+++ common-libs/tools-common/.env.example
@@ -1,0 +1,1 @@
+TEST_SECRETS={}
