+++ common-libs/tools-common/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ["@affinidi/eslint-config"]
+};
