+++ common-libs/tools-common/README.md
@@ -1,0 +1,5 @@
+# Affinidi common tools
+
+Affinidi specific utilities (`SdkError`, `@profile`), combined in a separate package for convenience. 
+
+Only affinidi projects are supposed to depend on this package.
