+++ common-libs/tools-common/src/SdkError.ts
