+++ common-libs/tools-common/src/profiler/ConsoleReporter.ts
