+++ common-libs/tools-common/src/profiler/DefaultProfilerActivator.ts
