+++ common-libs/tools-common/src/profiler/IProfileActivator.ts
