+++ common-libs/tools-common/test/unit/index.test.ts
@@ -1,0 +1,3 @@
+import 'mocha'
+
+require('./ProfilerDecorator.test')
