+++ common-libs/tools-common/tsconfig-build.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["node_modules", "test/**/*"]
+}
+
