+++ sdk/core/package.json
@@ -47,6 +47,7 @@
     "@affinidi/internal-api-clients": "^1.0.0-beta.10",
     "@affinidi/issuer-email-ses-client": "^0.1.1",
     "@affinidi/issuer-phone-twilio-client": "^0.1.1",
+    "@affinidi/tools-common": "^1.0.0-beta.1",
     "@affinidi/vc-common": "^1.4.2",
     "@affinidi/vc-data": "^1.3.3",
     "ajv": "^6.12.3",
