+++ sdk/core/src/services/KeyManagementService.ts
@@ -1,5 +1,6 @@
-import { JwtService, KeysService, profile } from '@affinidi/common'
+import { JwtService, KeysService } from '@affinidi/common'
 import { KeyStorageApiService } from '@affinidi/internal-api-clients'
+import { profile } from '@affinidi/tools-common'
 import retry from 'async-retry'
 import { extractSDKVersion } from '../_helpers'
 const createHash = require('create-hash')
