+++ sdk/core/test/integration/otp-legacy/core.test.ts
@@ -2,7 +2,7 @@ import 'mocha'
 import '../env'
 
 import { expect } from 'chai'
-import { SdkError } from '@affinidi/common'
+import { SdkError } from '@affinidi/tools-common'
 import { AffinidiWallet, checkIsWallet } from '../../helpers/AffinidiWallet'
 import { SdkOptions } from '../../../src/dto/shared.dto'
